Accelerate the second phase using burst
This isn't doing much work anyway, so it makes more sense to do it all
on a single thread instead of launching a bunch of jobs to do it.

This commit adds a regular versions using Unity.Mathematics as well as a
properly vectorized version for AVX2. The codegen for the mathematics
version isn't great but I don't think it's worth spending more effort
optimizing this since it's a pretty small part of the overall time.